Correcciones al modelo
se ha agregado el modelo guardados, se ha hecho la corrección de la tabla tipoPublicidad

+++ api/models/TipoPublicidad.js
@@ -17,7 +17,7 @@
        columnName: 'idTipoPublicidad'
      },
      costo: {
-       type: 'double',
+       type: 'float',
        columnName: 'costo'
      },
      duracion :{
